requirements.txt
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.24.3
python-chess==1.999